Refactored the Reorder common code to use primitive input types rather than Frame.
No changes to the reorder-frame sample code, just a makefile fixup.
